app/requirements.txt
gunicorn==20.0.4
flask==1.1.2
mysql-connector-python